refactored 'sendChangeRequest' and 'isDisabledForm'
